Start refactoring messaging and handshake code
