+++ go.mod
@@ -3,11 +3,14 @@ module websocket-chat
 go 1.23.4
 
 require (
-	github.com/fatih/color v1.18.0 // indirect
+	github.com/fatih/color v1.18.0
+	github.com/google/uuid v1.6.0
+	github.com/gorilla/websocket v1.5.3
+)
+
+require (
 	github.com/gdamore/encoding v1.0.0 // indirect
 	github.com/gdamore/tcell/v2 v2.7.1 // indirect
-	github.com/google/uuid v1.6.0 // indirect
-	github.com/gorilla/websocket v1.5.3 // indirect
 	github.com/lucasb-eyer/go-colorful v1.2.0 // indirect
 	github.com/mattn/go-colorable v0.1.13 // indirect
 	github.com/mattn/go-isatty v0.0.20 // indirect
