+++ util/util.go
@@ -11,12 +11,43 @@ import (
 	"errors"
 	"fmt"
 	"io"
+	"log"
 	"math/big"
+	"net/url"
+
+	"github.com/gorilla/websocket"
+)
+
+var (
+	P, G, privateKey, publicKey *big.Int
+	roomKey                     []byte
 )
 
 // Clear the current line in the terminal after pressing return
 func ClearLine() {
 	fmt.Print("\033[1A\033[K")
+}
+
+func ClearTerminal() {
+	fmt.Print("\033[H\033[2J")
+}
+
+func GenerateKeys() {
+	P = GeneratePrime()
+	G = big.NewInt(2)
+	GeneratePrivateKey()
+	CalculatePublicKey(G)
+	checkRoomKey()
+}
+
+func checkRoomKey() {
+	if roomKey == nil {
+		roomKey = make([]byte, 32)
+		_, err := rand.Read(roomKey)
+		if err != nil {
+			log.Println("room key:", err)
+		}
+	}
 }
 
 func Encrypt(plaintext []byte, key []byte) (string, error) {
@@ -69,19 +100,127 @@ func GeneratePrime() *big.Int {
 }
 
 // Both parties generate private keys
-func GeneratePrivateKey(prime *big.Int) *big.Int {
-	privateKey, _ := rand.Int(rand.Reader, prime)
-	return privateKey
+func GeneratePrivateKey() {
+	privateKey, _ = rand.Int(rand.Reader, P)
 }
 
 // Both parties calculate their own public keys
-func CalculatePublicKey(prime, privateKey, base *big.Int) *big.Int {
-	publicKey := new(big.Int).Exp(base, privateKey, prime)
-	return publicKey
+func CalculatePublicKey(base *big.Int) {
+	publicKey = new(big.Int).Exp(base, privateKey, P)
 }
 
 // Parties exchange public keys and use them to calculate shared secret
-func CalculateSharedSecret(prime, privateKey, publicKey *big.Int) *big.Int {
-	sharedSecret := new(big.Int).Exp(publicKey, privateKey, prime)
+func CalculateSharedSecret(publicKeyRemote *big.Int) *big.Int {
+	sharedSecret := new(big.Int).Exp(publicKeyRemote, privateKey, P)
 	return sharedSecret
+}
+
+func DoKeyExchange(conn *websocket.Conn) error {
+	// Receive P, G, key hub public key from server
+	_, Pbytes, err := conn.ReadMessage()
+	if err != nil {
+		newError := errors.New("Error receiving P from server:" + err.Error())
+		return newError
+	}
+	P = new(big.Int).SetBytes(Pbytes)
+
+	_, Gbytes, err := conn.ReadMessage()
+	if err != nil {
+		newError := errors.New("Error receiving G from server:" + err.Error())
+		return newError
+	}
+	G = new(big.Int).SetBytes(Gbytes)
+
+	_, keyHubPubKeyBytes, err := conn.ReadMessage()
+	if err != nil {
+		newError := errors.New("Error receiving key hub's public key:" + err.Error())
+		return newError
+	}
+	keyHubPubKey := new(big.Int).SetBytes(keyHubPubKeyBytes)
+
+	// Calculate private key
+	GeneratePrivateKey()
+
+	// Calculate public key
+	CalculatePublicKey(G)
+
+	// Send public key to server
+	err = conn.WriteMessage(websocket.BinaryMessage, publicKey.Bytes())
+	if err != nil {
+		newError := errors.New("Error sending public key to server:" + err.Error())
+		return newError
+	}
+	// Calculate PSK with server pub key
+	psk := CalculateSharedSecret(keyHubPubKey)
+
+	// Receive room key from server
+	_, encryptedRoomKeyBytes, err := conn.ReadMessage()
+	if err != nil {
+		newError := errors.New("Error receiving room key from server:" + err.Error())
+		return newError
+	}
+
+	roomKey, err = Decrypt(string(encryptedRoomKeyBytes), psk.Bytes())
+	if err != nil {
+		newError := errors.New("Error decrypting room key:" + err.Error())
+		return newError
+	}
+	return nil
+}
+
+func ShareKeys(hostName *string, hostPort *int) error {
+	u := url.URL{Scheme: "ws", Host: fmt.Sprintf("%s:%d", *hostName, *hostPort), Path: "/key-exchange"}
+	conn, _, err := websocket.DefaultDialer.Dial(u.String(), nil)
+	if err != nil {
+		log.Fatal(err)
+		return err
+	}
+	defer conn.Close()
+
+	// Send P, G, public key to server
+	err = conn.WriteMessage(websocket.BinaryMessage, P.Bytes())
+	if err != nil {
+		newError := errors.New("Error sending P to server:" + err.Error())
+		return newError
+	}
+
+	err = conn.WriteMessage(websocket.BinaryMessage, G.Bytes())
+	if err != nil {
+		newError := errors.New("Error sending G to server:" + err.Error())
+		return newError
+	}
+
+	err = conn.WriteMessage(websocket.BinaryMessage, publicKey.Bytes())
+	if err != nil {
+		newError := errors.New("Error sending public key to server:" + err.Error())
+		return newError
+	}
+
+	// Receive other client's public key
+	_, clientPubKeyBytes, err := conn.ReadMessage()
+	if err != nil {
+		newError := errors.New("Error receiving client's public key:" + err.Error())
+		return newError
+	}
+	clientPubKey := new(big.Int).SetBytes(clientPubKeyBytes)
+
+	// Calculate PSK
+	psk := CalculateSharedSecret(clientPubKey)
+
+	// Send encrypted room key
+	encryptedRoomKey, err := Encrypt(roomKey, psk.Bytes())
+	if err != nil {
+		newError := errors.New("Error encrypting room key:" + err.Error())
+		return newError
+	}
+	err = conn.WriteMessage(websocket.BinaryMessage, []byte(encryptedRoomKey))
+	if err != nil {
+		newError := errors.New("Error sending room key to server:" + err.Error())
+		return newError
+	}
+	return nil
+}
+
+func GetRoomKey() []byte {
+	return roomKey
 }
